Added utils module
